Course new dialog now has container and is presentational
